requirements.txt
Django==5.1
celery==5.4.0
djangorestframework==3.15.2
django-celery-beat==2.7.0
psycopg2-binary==2.9.9
redis==5.0.8
python-dotenv==1.0.1